requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
pandas==1.4.3
plotly==5.10.0
psycopg2-binary==2.9.3
streamlit==1.12.2
config
